Update ratings API v2
